[+] Common String constructor for all platforms
Previously `TCHAR` was `wchar_t` on all platforms. Recently it was changed to `char16_t` on Android. This meant that all `AString` constructor calls failed to compile for the platform. This change creates a new constructor that takes a `TCHAR` and removes the now redundant `wchar_t` version so that Win platforms compile.

Also anywhere we were previously calling `TCHAR_TO_WCHAR` goes away in favor of the `TCHAR`.