[~] Do not generate types for Skipped Classes children. Historically the behavior of SkipClasses in the config file has been that only the direct class specified by the `+SkipClasses` entry was skipped. This meant that for classes A and B, if we skipped class A and class B inherited from A, class B would still be included. This was accomplished by creating stub entries for any skipped classes.
One challenge this produced was when trying to skip large swaths of reflected engine classes, particularly animation graph related classes.
These are classes with deep class hierarchies that when used in a normal project can have several duplicate variable names that end up
confusing and breaking SkookumScript.

My thoughts were that I could add 7 new classes to the default `SkipClasses` to prevent others from hitting the bug. But in the next release,
Epic could add new subclasses and we'd run the risk of encountering the issue again. And this could be a silent trigger because you'd need
the right project that utilized these new classes and generated the workload that creates the duplicate node names.

It seemed that the better solution was to ignore everything below a specified `+SkipClasses` entry so that subclasses of the skipped class are
automatically culled. That is what this change does.